Separated the feature for displaying items in a tabular layout from the Box type, and introduced a new Grid type under the grid subpackage.
