+++ box.go
@@ -2,24 +2,22 @@ package box
 
 import (
 	"bufio"
-	"bytes"
 	"context"
 	"fmt"
 	"io"
 	"sort"
-	"strings"
 
 	"github.com/mattn/go-tty"
+
+	"github.com/nyaosorg/go-box/v3/grid"
 
 	"github.com/nyaosorg/go-box/v3/internal/ansi"
 	"github.com/nyaosorg/go-box/v3/internal/keys"
 )
 
 type Box struct {
-	width  int
-	height int
-	cache  [][]byte
-	tty    _Tty
+	grid.Grid
+	tty _Tty
 }
 
 func New() (*Box, error) {
@@ -29,112 +27,16 @@ func New() (*Box, error) {
 	}
 	w, h, err := tty1.Size()
 	return &Box{
-		width:  w,
-		height: h,
-		tty:    _GoTty{TTY: tty1},
+		Grid: grid.Grid{
+			Width:  w,
+			Height: h,
+		},
+		tty: _GoTty{TTY: tty1},
 	}, err
 }
 
 func (b *Box) Close() error {
 	return b.tty.Close()
-}
-
-// Println outputs the given items in a tabular layout and appends
-// a newline after the final line.
-func (b *Box) Println(
-	nodes []string,
-	offset int,
-	out io.Writer) (int, int, error) {
-
-	columns, nlines, err := b.Print(nodes, offset, out)
-	if err != nil {
-		return columns, nlines, err
-	}
-	// append last linefeed.
-	if nlines > 0 {
-		fmt.Fprintln(out)
-	}
-	b.cache = nil
-	return columns, nlines, nil
-}
-
-// Print outputs the given items in a tabular layout without appending
-// a trailing newline. It returns the number of rows and columns printed.
-func (b *Box) Print(
-	nodes []string,
-	offset int,
-	out io.Writer) (int, int, error) {
-
-	if len(nodes) <= 0 {
-		return 0, 0, nil
-	}
-
-	rw := wtRuneWidth.Value()
-	maxLen := 1
-	for _, finfo := range nodes {
-		length := rw.StringWidth(ansi.RxSequence.ReplaceAllString(finfo, ""))
-		if length > maxLen {
-			maxLen = length
-		}
-	}
-	nodePerLine := (b.width - 1) / (maxLen + 1)
-	if nodePerLine <= 0 {
-		nodePerLine = 1
-	}
-	nlines := (len(nodes) + nodePerLine - 1) / nodePerLine
-
-	lines := make([][]byte, nlines)
-	row := 0
-	for _, finfo := range nodes {
-		lines[row] = append(lines[row], finfo...)
-		w := rw.StringWidth(ansi.RxSequence.ReplaceAllString(finfo, ""))
-		if maxLen < b.width {
-			for i := maxLen + 1; i > w; i-- {
-				lines[row] = append(lines[row], ' ')
-			}
-		}
-		row++
-		if row >= nlines {
-			row = 0
-		}
-	}
-	i_end := len(lines)
-	if b.height > 0 {
-		if i_end >= offset+b.height {
-			i_end = offset + b.height
-		}
-	}
-
-	if b.cache == nil {
-		b.cache = make([][]byte, b.height)
-	}
-	i := offset
-	y := 0
-	for {
-		if y >= len(b.cache) {
-			b.cache = append(b.cache, []byte{})
-		}
-		// assertion
-		if i >= len(lines) {
-			return 0, 0, fmt.Errorf("assertion failed: len(lines)==%d i==%d", len(lines), i)
-		}
-		if !bytes.Equal(lines[i], b.cache[y]) {
-			line := strings.TrimRight(string(lines[i]), " ")
-			line = cutReduntantColorChange(line)
-			io.WriteString(out, line)
-			if len(b.cache[y]) > 0 {
-				fmt.Fprint(out, ansi.EraseLine)
-			}
-			b.cache[y] = lines[i]
-		}
-		y++
-		i++
-		if i >= i_end {
-			break
-		}
-		fmt.Fprintln(out)
-	}
-	return nodePerLine, nlines, nil
 }
 
 // SelectIndex returns the indexes that user selected.
@@ -149,7 +51,7 @@ func (b *Box) SelectIndexContext(ctx context.Context, sources []string, multi bo
 	nodes := make([]nodeT, 0, len(sources))
 	draws := make([]string, 0, len(sources))
 	for i, text := range sources {
-		val := truncate(text, b.width-1)
+		val := grid.Truncate(text, b.Width-1)
 		if val != "" {
 			nodes = append(nodes, nodeT{Index: i, Text: val})
 			draws = append(draws, val)
@@ -166,17 +68,17 @@ func (b *Box) SelectIndexContext(ctx context.Context, sources []string, multi bo
 	offset := 0
 	for {
 		for index := range selected {
-			draws[index] = ansi.BoldOn + truncate(nodes[index].Text, b.width-2) + ansi.BoldOff
+			draws[index] = ansi.BoldOn + grid.Truncate(nodes[index].Text, b.Width-2) + ansi.BoldOff
 		}
-		draws[cursor] = ansi.BoldOn2 + truncate(nodes[cursor].Text, b.width-2) + ansi.BoldOff
+		draws[cursor] = ansi.BoldOn2 + grid.Truncate(nodes[cursor].Text, b.Width-2) + ansi.BoldOff
 		_, h, err := b.Print(draws, offset, out)
 		if err != nil {
 			return []int{}, err
 		}
 		for index := range selected {
-			draws[index] = truncate(nodes[index].Text, b.width-2)
+			draws[index] = grid.Truncate(nodes[index].Text, b.Width-2)
 		}
-		draws[cursor] = truncate(nodes[cursor].Text, b.width-2)
+		draws[cursor] = grid.Truncate(nodes[cursor].Text, b.Width-2)
 		last := cursor
 
 		var doSelect func()
@@ -242,18 +144,18 @@ func (b *Box) SelectIndexContext(ctx context.Context, sources []string, multi bo
 			if y < offset {
 				offset = y
 				// offset--
-			} else if y >= offset+b.height {
-				offset = y - b.height + 1
+			} else if y >= offset+b.Height {
+				offset = y - b.Height + 1
 				// offset++
 			}
 		}
-		if h < b.height {
+		if h < b.Height {
 			if h > 1 {
 				fmt.Fprintf(out, ansi.UpN, h-1)
 			}
 		} else {
-			if b.height > 1 {
-				fmt.Fprintf(out, ansi.UpN, b.height-1)
+			if b.Height > 1 {
+				fmt.Fprintf(out, ansi.UpN, b.Height-1)
 			}
 		}
 		fmt.Fprint(out, "\r")
